fastcomp: fix integer overflow
If the pattern length exceeds INT_MAX, the length will be cast to an
signed int:

util.c:563:13: warning: Call to 'malloc' has an allocation size of 0 bytes
        if ((ptr = malloc(size)) == NULL)
                   ^~~~~~~~~~~~

+++ grep.h
@@ -54,7 +54,7 @@ typedef struct {
 
 typedef struct {
 	unsigned char	*pattern;
-	int		 patternLen;
+	size_t	 patternLen;
 	int		 qsBc[UCHAR_MAX + 1];
 	/* flags */
 	int		 bol;
